[refactor] refactors iceboxList into functional component and fixes require statements
