+++ src/containers/visibleIceboxList.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import IceboxList from './iceboxList';
+import IceboxList from '../components/iceboxList';
 import { SORT_EXPIRATION, SORT_FOODGROUP, SORT_FOODNAME, ASCENDING, DESCENDING } from '../constants/sorts';
 import _sortBy from 'lodash/collection/sortBy';
